laravel_folder_structure: Modified Controller Folder Structure and moved try-catch block from Service class to the Controller
